FROMLIST: net: stmmac: dwmac-qcom-ethqos: Mask PHY mode if configured with rgmii-id
The Qualcomm board always chooses the MAC to provide the delay instead of
the PHY, which is completely opposite to the suggestion of the Linux
kernel. The usage of phy-mode in legacy DTS was also incorrect. Change the
phy_mode passed from the DTS to the driver from PHY_INTERFACE_MODE_RGMII_ID
to PHY_INTERFACE_MODE_RGMII to ensure correct operation and adherence to
the definition.
To address the ABI compatibility issue between the kernel and DTS caused by
this change, handle the compatible string 'qcom,qcs404-evb-4000' in the
code, as it is the only legacy board that mistakenly uses the 'rgmii'
phy-mode.

Link: https://lore.kernel.org/all/20250121-dts_qcs615-v3-2-fa4496950d8a@quicinc.com/
Signed-off-by: Yijie Yang <quic_yijiyang@quicinc.com>